+++ ChannelIO/Source/Reducers/GuestReducer.swift
@@ -10,14 +10,14 @@ import ReSwift
 
 func guestReducer(action: Action, guest: CHGuest?) -> CHGuest {
   switch action {
-  case let action as CheckInSuccess:
-    if let key = action.payload["guestKey"] as? String {
+  case let action as BootSuccess:
+    if let key = action.payload.guestKey as? String {
       PrefStore.setCurrentGuestKey(key)
     }
-    if let user = action.payload["user"] as? CHUser {
+    if let user = action.payload.user {
       PrefStore.setCurrentUserId(userId: user.id)
       return user
-    } else if let veil = action.payload["veil"] as? CHVeil {
+    } else if let veil = action.payload.veil {
       PrefStore.setCurrentVeilId(veilId: veil.id)
       return veil
     }
@@ -49,15 +49,16 @@ func guestReducer(action: Action, guest: CHGuest?) -> CHGuest {
     return guest ?? CHVeil()
     
   case let action as DeleteUserChat:
-    if let session = action.payload.session,
-      var guest = guest, let alert = guest.alert {
-      let adjustcount = alert - session.alert
-      guest.alert = adjustcount
+    if action.payload.isLocal,
+      let session = action.payload.session,
+      var guest = guest,
+      let alert = guest.alert {
+      guest.alert = alert - session.alert
       return guest
     }
     return guest ?? CHVeil()
   
-  case _ as CheckOutSuccess:
+  case _ as ShutdownSuccess:
     return CHVeil()
     
   default: